Upload the SingleGame.java

+++ Four_in_a_Line_v2/src/GameMain/Frame/GameKeyListener.java
@@ -26,6 +26,6 @@ public interface GameKeyListener {
      * want to place his piece.
      * @param keyInput The position you want to place then would send to the server.
      */
-    void sendKeyInput(int keyInput);
+    void sendKeyInput(byte keyInput);
 
 }
